Consolidate and rename Pub/Sub samples
