+++ pubsub/api/web/global.json
@@ -1,0 +1,5 @@
+﻿{
+  "sdk": {
+    "version": "1.0.0-beta8"
+  }
+}
